Move SimDoublets to SimDataFormats/TrackingAnalysis and SimDoubletsProducer to SimTracker/TrackerHitAssociation
