+++ RecoTracker/PixelSeeding/BuildFile.xml
@@ -23,7 +23,6 @@
 <use name="RecoTracker/TkMSParametrization"/>
 <use name="RecoTracker/TkTrackingRegions"/>
 <use name="TrackingTools/DetLayers"/>
-<use name="SimTracker/TrackerHitAssociation"/>
 <use name="vdt_headers"/>
 <export>
   <lib name="1"/>
