+++ RecoTracker/PixelSeeding/src/classes.h
@@ -1,5 +1,4 @@
 #include "RecoTracker/PixelSeeding/interface/IntermediateHitTriplets.h"
-#include "RecoTracker/PixelSeeding/interface/SimDoublets.h"
 #include "DataFormats/Common/interface/Wrapper.h"
 
 #include <vector>
@@ -11,11 +10,3 @@ namespace RecoPixelVertexing_PixelTriplets {
   };
 }  // namespace RecoPixelVertexing_PixelTriplets
 
-
-namespace {
-  struct simDoubletsDictionary {
-    SimDoublets sb;
-    edm::Wrapper<SimDoublets> wsb;
-    SiPixelRecHitRefVector sprhrv;
-  };
-}  // namespace
