+++ SimDataFormats/TrackingAnalysis/BuildFile.xml
@@ -6,6 +6,7 @@
 <use name="SimDataFormats/GeneratorProducts"/>
 <use name="SimDataFormats/Track"/>
 <use name="SimDataFormats/Vertex"/>
+<use name="Geometry/TrackerGeometryBuilder"/>
 <use name="FWCore/MessageLogger"/>
 <use name="clhepheader"/>
 <export>
