+++ SimDataFormats/TrackingAnalysis/src/classes_def.xml
@@ -40,4 +40,9 @@
   <class name="edm::Wrapper<std::unordered_map<UniqueSimTrackId, TrackingParticleRef, UniqueSimTrackIdHash>>"/>
   <class name="SimTrackToTPMap"/>
   <class name="edm::Wrapper<SimTrackToTPMap>"/>
+
+  <class name="SimDoublets"/>
+  <class name="std::vector<SimDoublets>"/>
+  <class name="edm::Wrapper< std::vector<SimDoublets> >"/>
+  <class name="SiPixelRecHitRefVector"/>
 </lcgdict>
