+++ SimTracker/TrackerHitAssociation/BuildFile.xml
@@ -1,6 +1,7 @@
 <use name="FWCore/Framework"/>
 <use name="FWCore/Utilities"/>
 <use name="DataFormats/Common"/>
+<use name="Geometry/Records"/>
 <use name="SimDataFormats/CrossingFrame"/>
 <use name="SimDataFormats/TrackingHit"/>
 <use name="SimDataFormats/TrackerDigiSimLink"/>
